backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==13.0
python-dotenv==1.0.1
# New unified Google GenAI SDK (replaces deprecated google-generativeai)
google-genai>=1.0.0
google-cloud-firestore==2.16.0
googlemaps==4.10.0
pydantic==2.7.1
httpx==0.28.1
pillow==10.3.0
python-multipart==0.0.9
aiofiles==23.2.1
